// src/typings/configurationWebhooks/amount.ts
/*
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */


export class Amount {
    /**
    * The three-character [ISO currency code](https://docs.adyen.com/development-resources/currency-codes#currency-codes).
    */
    "currency": string;
    /**
    * The amount of the transaction, in [minor units](https://docs.adyen.com/development-resources/currency-codes#minor-units).
    */
    "value": number;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "currency",
            "baseName": "currency",
            "type": "string",
            "format": ""
        },
        {
            "name": "value",
            "baseName": "value",
            "type": "number",
            "format": "int64"
        }    ];

    static getAttributeTypeMap() {
        return Amount.attributeTypeMap;
    }

    public constructor() {
    }
}