// src/typings/configurationWebhooks/balancePlatformNotificationResponse.ts
/*
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */


export class BalancePlatformNotificationResponse {
    /**
    * Respond with any **2xx** HTTP status code to [accept the webhook](https://docs.adyen.com/development-resources/webhooks#accept-notifications).
    */
    "notificationResponse"?: string;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "notificationResponse",
            "baseName": "notificationResponse",
            "type": "string",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return BalancePlatformNotificationResponse.attributeTypeMap;
    }

    public constructor() {
    }
}