// src/typings/configurationWebhooks/capabilityProblem.ts
/*
 * The version of the OpenAPI document: v2
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */

import { CapabilityProblemEntity } from "./capabilityProblemEntity";
import { VerificationError } from "./verificationError";


export class CapabilityProblem {
    "entity"?: CapabilityProblemEntity;
    /**
    * Contains information about the verification error.
    */
    "verificationErrors"?: Array<VerificationError>;

    static readonly discriminator: string | undefined = undefined;

    static readonly mapping: {[index: string]: string} | undefined = undefined;

    static readonly attributeTypeMap: Array<{name: string, baseName: string, type: string, format: string}> = [
        {
            "name": "entity",
            "baseName": "entity",
            "type": "CapabilityProblemEntity",
            "format": ""
        },
        {
            "name": "verificationErrors",
            "baseName": "verificationErrors",
            "type": "Array<VerificationError>",
            "format": ""
        }    ];

    static getAttributeTypeMap() {
        return CapabilityProblem.attributeTypeMap;
    }

    public constructor() {
    }
}