// src/typings/configurationWebhooks/expiry.ts
/*
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */


export class Expiry {
    /**
    * The month in which the card will expire.
    */
    'month'?: string;
    /**
    * The year in which the card will expire.
    */
    'year'?: string;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "month",
            "baseName": "month",
            "type": "string"
        },
        {
            "name": "year",
            "baseName": "year",
            "type": "string"
        }    ];

    static getAttributeTypeMap() {
        return Expiry.attributeTypeMap;
    }
}