// src/typings/configurationWebhooks/phone.ts
/*
 * The version of the OpenAPI document: v1
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit this class manually.
 */


export class Phone {
    /**
    * The full phone number provided as a single string.  For example, **\"0031 6 11 22 33 44\"**, **\"+316/1122-3344\"**,    or **\"(0031) 611223344\"**.
    */
    'number': string;
    /**
    * Type of phone number. Possible values:  **Landline**, **Mobile**. 
    */
    'type': Phone.TypeEnum;

    static discriminator: string | undefined = undefined;

    static attributeTypeMap: Array<{name: string, baseName: string, type: string}> = [
        {
            "name": "number",
            "baseName": "number",
            "type": "string"
        },
        {
            "name": "type",
            "baseName": "type",
            "type": "Phone.TypeEnum"
        }    ];

    static getAttributeTypeMap() {
        return Phone.attributeTypeMap;
    }
}

export namespace Phone {
    export enum TypeEnum {
        Landline = 'Landline',
        Mobile = 'Mobile'
    }
}